+++ apps/oidnBenchmark.cpp
@@ -250,12 +250,12 @@ int main(int argc, char* argv[])
         if (width < 1 || height < 1)
           throw std::runtime_error("invalid image size");
       }
-      else if (opt == "t" || opt == "type" || opt == "dtype")
+      else if (opt == "t" || opt == "type")
       {
         const auto val = args.getNextValue();
-        if (val == "float" || val == "Float" || val == "fp32" || val == "f32" || val == "f")
+        if (val == "f" || val == "float" || val == "Float" || val == "fp32")
           dataType = Format::Float;
-        else if (val == "half" || val == "Half" || val == "fp16" || val == "f16" || val == "h")
+        else if (val == "h" || val == "half" || val == "Half" || val == "fp16")
           dataType = Format::Half;
         else
           throw std::runtime_error("invalid data type");
@@ -301,6 +301,9 @@ int main(int argc, char* argv[])
     // Initialize the device
     DeviceRef device = newDevice(deviceType);
 
+    if (verbose >= 0)
+      device.set("verbose", verbose);
+
     const char* errorMessage;
     if (device.getError(errorMessage) != Error::None)
       throw std::runtime_error(errorMessage);
@@ -310,8 +313,7 @@ int main(int argc, char* argv[])
       device.set("numThreads", numThreads);
     if (setAffinity >= 0)
       device.set("setAffinity", bool(setAffinity));
-    if (verbose >= 0)
-      device.set("verbose", verbose);
+
     device.commit();
 
     // Run the benchmarks