+++ apps/oidnTest.cpp
@@ -32,11 +32,11 @@ void setFilterImage(FilterRef& filter, const char* name, ImageBuffer& image)
   filter.setImage(name, image.data(), format, image.width, image.height);
 }
 
-ImageBuffer makeConstImage(DeviceRef& device, int W, int H, int C = 3, float value = 0.5f)
+std::shared_ptr<ImageBuffer> makeConstImage(DeviceRef& device, int W, int H, int C = 3, float value = 0.5f)
 {
-  ImageBuffer image(device, W, H, C);
-  for (size_t i = 0; i < image.size(); ++i)
-    image.set(i, value);
+  auto image = std::make_shared<ImageBuffer>(device, W, H, C);
+  for (size_t i = 0; i < image->size(); ++i)
+    image->set(i, value);
   return image;
 }
 
@@ -59,16 +59,16 @@ TEST_CASE("single filter", "[single_filter]")
   const int H = 89;
 
   DeviceRef device = newDevice();
-  REQUIRE(device);
+  REQUIRE(bool(device));
   device.commit();
   REQUIRE(device.getError() == Error::None);
 
   FilterRef filter = device.newFilter("RT");
-  REQUIRE(filter);
+  REQUIRE(bool(filter));
 
-  ImageBuffer image = makeConstImage(device, W, H);
-  setFilterImage(filter, "color",  image);
-  setFilterImage(filter, "output", image);
+  std::shared_ptr<ImageBuffer> image = makeConstImage(device, W, H);
+  setFilterImage(filter, "color",  *image);
+  setFilterImage(filter, "output", *image);
 
   filter.commit();
   REQUIRE(device.getError() == Error::None);
@@ -96,11 +96,11 @@ void multiFilter1PerDeviceTest(DeviceRef& device, const std::vector<int>& sizes)
   for (size_t i = 0; i < sizes.size(); ++i)
   {
     FilterRef filter = device.newFilter("RT");
-    REQUIRE(filter);
+    REQUIRE(bool(filter));
 
-    ImageBuffer image = makeConstImage(device, sizes[i], sizes[i]);
-    setFilterImage(filter, "color",  image);
-    setFilterImage(filter, "output", image);
+    std::shared_ptr<ImageBuffer> image = makeConstImage(device, sizes[i], sizes[i]);
+    setFilterImage(filter, "color",  *image);
+    setFilterImage(filter, "output", *image);
 
     filter.commit();
     REQUIRE(device.getError() == Error::None);
@@ -113,16 +113,16 @@ void multiFilter1PerDeviceTest(DeviceRef& device, const std::vector<int>& sizes)
 void multiFilterNPerDeviceTest(DeviceRef& device, const std::vector<int>& sizes)
 {
   std::vector<FilterRef> filters;
-  std::vector<ImageBuffer> images;
+  std::vector<std::shared_ptr<ImageBuffer>> images;
 
   for (size_t i = 0; i < sizes.size(); ++i)
   {
     filters.push_back(device.newFilter("RT"));
-    REQUIRE(filters[i]);
+    REQUIRE(bool(filters[i]));
 
     images.push_back(makeConstImage(device, sizes[i], sizes[i]));
-    setFilterImage(filters[i], "color",  images[i]);
-    setFilterImage(filters[i], "output", images[i]);
+    setFilterImage(filters[i], "color",  *images[i]);
+    setFilterImage(filters[i], "output", *images[i]);
 
     filters[i].commit();
     REQUIRE(device.getError() == Error::None);
@@ -138,7 +138,7 @@ void multiFilterNPerDeviceTest(DeviceRef& device, const std::vector<int>& sizes)
 TEST_CASE("multiple filters", "[multi_filter]")
 {
   DeviceRef device = newDevice();
-  REQUIRE(device);
+  REQUIRE(bool(device));
   device.commit();
   REQUIRE(device.getError() == Error::None);
 
@@ -171,7 +171,7 @@ TEST_CASE("multiple devices", "[multi_device]")
 
   std::vector<DeviceRef> devices;
   std::vector<FilterRef> filters;
-  std::vector<ImageBuffer> images;
+  std::vector<std::shared_ptr<ImageBuffer>> images;
 
   for (size_t i = 0; i < sizes.size(); ++i)
   {
@@ -181,11 +181,11 @@ TEST_CASE("multiple devices", "[multi_device]")
     REQUIRE(devices[i].getError() == Error::None);
 
     filters.push_back(devices[i].newFilter("RT"));
-    REQUIRE(filters[i]);
+    REQUIRE(bool(filters[i]));
 
     images.push_back(makeConstImage(devices[i], sizes[i], sizes[i]));
-    setFilterImage(filters[i], "color",  images[i]);
-    setFilterImage(filters[i], "output", images[i]);
+    setFilterImage(filters[i], "color",  *images[i]);
+    setFilterImage(filters[i], "output", *images[i]);
 
     filters[i].commit();
     REQUIRE(devices[i].getError() == Error::None);
@@ -206,19 +206,19 @@ TEST_CASE("filter update", "[filter_update]")
   const int H = 599;
 
   DeviceRef device = newDevice();
-  REQUIRE(device);
+  REQUIRE(bool(device));
   device.commit();
   REQUIRE(device.getError() == Error::None);
 
   FilterRef filter = device.newFilter("RT");
-  REQUIRE(filter);
+  REQUIRE(bool(filter));
 
-  ImageBuffer color  = makeConstImage(device, W, H);
-  ImageBuffer albedo = makeConstImage(device, W, H);
-  ImageBuffer output = makeConstImage(device, W, H);
-  setFilterImage(filter, "color",  color);
-  setFilterImage(filter, "albedo", albedo);
-  setFilterImage(filter, "output", output);
+  std::shared_ptr<ImageBuffer> color  = makeConstImage(device, W, H);
+  std::shared_ptr<ImageBuffer> albedo = makeConstImage(device, W, H);
+  std::shared_ptr<ImageBuffer> output = makeConstImage(device, W, H);
+  setFilterImage(filter, "color",  *color);
+  setFilterImage(filter, "albedo", *albedo);
+  setFilterImage(filter, "output", *output);
 
   filter.set("hdr", true);
 
@@ -236,7 +236,7 @@ TEST_CASE("filter update", "[filter_update]")
   SECTION("filter update: same image size")
   {
     color = makeConstImage(device, W, H);
-    setFilterImage(filter, "color", color);
+    setFilterImage(filter, "color", *color);
   }
 
   SECTION("filter update: different image size")
@@ -244,9 +244,9 @@ TEST_CASE("filter update", "[filter_update]")
     color  = makeConstImage(device, W*2, H*2);
     albedo = makeConstImage(device, W*2, H*2);
     output = makeConstImage(device, W*2, H*2);
-    setFilterImage(filter, "color",  color);
-    setFilterImage(filter, "albedo", albedo);
-    setFilterImage(filter, "output", output);
+    setFilterImage(filter, "color",  *color);
+    setFilterImage(filter, "albedo", *albedo);
+    setFilterImage(filter, "output", *output);
   }
 
   SECTION("filter update: remove image")
@@ -276,7 +276,7 @@ TEST_CASE("filter update", "[filter_update]")
 void imageSizeTest(DeviceRef& device, int W, int H)
 {
   FilterRef filter = device.newFilter("RT");
-  REQUIRE(filter);
+  REQUIRE(bool(filter));
 
   const int N = std::max(W * H * 3, 1); // make sure the buffers are never null
   std::vector<float> input(N, 0.5f);
@@ -295,7 +295,7 @@ void imageSizeTest(DeviceRef& device, int W, int H)
 TEST_CASE("image size", "[size]")
 {
   DeviceRef device = newDevice();
-  REQUIRE(device);
+  REQUIRE(bool(device));
   device.commit();
   REQUIRE(device.getError() == Error::None);
 
@@ -327,13 +327,13 @@ void sanitizationTest(DeviceRef& device, bool hdr, float value)
   const int H = 347;
 
   FilterRef filter = device.newFilter("RT");
-  REQUIRE(filter);
+  REQUIRE(bool(filter));
 
-  ImageBuffer input = makeConstImage(device, W, H, 3, value);
+  std::shared_ptr<ImageBuffer> input = makeConstImage(device, W, H, 3, value);
   ImageBuffer output(device, W, H, 3);
-  setFilterImage(filter, "color",  input);
-  setFilterImage(filter, "albedo", input);
-  setFilterImage(filter, "normal", input);
+  setFilterImage(filter, "color",  *input);
+  setFilterImage(filter, "albedo", *input);
+  setFilterImage(filter, "normal", *input);
   setFilterImage(filter, "output", output);
   filter.set("hdr", hdr);
 
@@ -352,7 +352,7 @@ void sanitizationTest(DeviceRef& device, bool hdr, float value)
 TEST_CASE("image sanitization", "[sanitization]")
 {
   DeviceRef device = newDevice();
-  REQUIRE(device);
+  REQUIRE(bool(device));
   device.commit();
   REQUIRE(device.getError() == Error::None);
 
@@ -399,11 +399,11 @@ void progressTest(DeviceRef& device, double nMax = 1000)
   const int H = 727;
 
   FilterRef filter = device.newFilter("RT");
-  REQUIRE(filter);
+  REQUIRE(bool(filter));
 
-  ImageBuffer image = makeConstImage(device, W, H);
-  setFilterImage(filter, "color",  image);
-  setFilterImage(filter, "output", image); // in-place
+  std::shared_ptr<ImageBuffer> image = makeConstImage(device, W, H);
+  setFilterImage(filter, "color",  *image);
+  setFilterImage(filter, "output", *image); // in-place
 
   Progress progress(nMax);
   filter.setProgressMonitorFunction(progressCallback, &progress);
@@ -432,7 +432,7 @@ void progressTest(DeviceRef& device, double nMax = 1000)
 TEST_CASE("progress monitor", "[progress]")
 {
   DeviceRef device = newDevice();
-  REQUIRE(device);
+  REQUIRE(bool(device));
   device.commit();
   REQUIRE(device.getError() == Error::None);